style: refine explore page typography
