+++ src/components/exploreCommandInput/index.module.css
@@ -22,6 +22,8 @@
 
 .prompt {
   white-space: nowrap;
+  color: var(--foreground);
+  font-weight: bold;
 }
 
 .group {
