+++ src/components/repositories/grid/index.module.css
@@ -3,3 +3,8 @@
   grid-template-columns: repeat(auto-fit, minmax(min(100%, 24rem), 1fr));
   gap: var(--spacing-sm);
 }
+
+.homepageCard {
+  --card-border-width: 2px;
+  --card-border-color: var(--foreground-accent);
+}
